move test_flat_open_closed_icosmo
